fix(mobile): harden startup subscription lifecycle
Publish the bounded channel snapshot independently of live EOSE, quarantine
terminally closed channel subscriptions until a material invalidator, and
extract channel sync and relay support code below the file-size ceiling.
