+++ mobile/lib/shared/relay/relay.dart
@@ -9,6 +9,8 @@ export 'relay_closed_policy.dart';
 export 'relay_client.dart';
 export 'relay_provider.dart';
 export 'relay_rate_limit_gate.dart';
+export 'relay_nip98_auth.dart';
 export 'relay_session.dart';
+export 'relay_session_state.dart';
 export 'relay_socket.dart';
 export 'signed_event_relay.dart';
