+++ mobile/lib/shared/relay/relay_session_state.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/foundation.dart';
+
+enum SessionStatus { disconnected, connecting, connected, reconnecting }
+
+@immutable
+class SessionState {
+  final SessionStatus status;
+  final int reconnectAttempt;
+
+  const SessionState({required this.status, this.reconnectAttempt = 0});
+}
